requirements.txt
Django~=3.1.4
django-rest-auth~=0.9.5
djangorestframework~=3.12.2
djangorestframework-jwt~=1.11.0
django-allauth~=0.44.0